requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
lxml==5.2.2
modules==1.0.0
numpy==1.26.4
packaging==24.1
pandas
psycopg-binary==3.1.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
tzdata==2024.1
urllib3==2.2.2
